# rap-lp-openapi.yaml
...
openapi: 3.1.0
info:
  title: RAP-LP API
  version: 1.0.11
  summary: API for validating OpenAPI specifications against the Swedish REST API profile.
  description: |
    RAP-LP (REST API-profil Lint Processor) validates an OpenAPI specification against
    the national REST API profile and additional RAP-LP design rules.

    RAP-LP validates structure, semantics and design rules, and can also generate an
    Excel report from a precomputed validation report.

    This specification describes the linter API itself.

    Begransningar och problem: URL-baserad validering kan vara avstangd via serverkonfiguration,
    och generate-report returnerar en binar Excel-fil medan ovriga resurser returnerar JSON.
  contact:
    name: Digg
    url: https://github.com/diggsweden/rest-api-profil-lint-processor/
    email: apiteam@digg.se
  license:
    name: EUPL-1.2
    url: https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
  termsOfService: https://www.dataportal.se/terms
...
servers:
  - url: https://raplp.digg.se/api/v1/
    description: Production
  - url: https://raplp.sandbox.digg.se/api/v1/
    description: Sandbox
  - url: http://localhost:3000/api/v1/
    description: Local development
...
security: []
paths:
  /validation/rules:
    get:
      tags:
        - metadata
      summary: List available rule categories.
      description: Returns the available rule categories that can be supplied when validating a specification.
      operationId: listValidationRules
      responses:
        '200':
          description: Available rule categories were returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationRules'
              examples:
                default:
                  value:
                    - rule: UfnRules
                      description: URL format och namngivning
                    - rule: DokRules
                      description: Dokumentation
                    - rule: AmeRules
                      description: API Message
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                internalServerError:
                  value:
                    type: https://raplp.digg.se/problems/internal-server-error
                    title: Internt serverfel
                    status: 500
                    detail: Ett ovantat fel intraffade.
                    instance: /api/v1/validation/rules
  /api-info:
    get:
      tags:
        - metadata
      summary: Get API information.
      description: Returns name, semantic version, release date, documentation URL and lifecycle status for the RAP-LP API.
      operationId: getApiInfo
      responses:
        '200':
          description: API information was returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiInfo'
              examples:
                default:
                  value:
                    apiName: RAP-LP
                    apiVersion: 1.0.11
                    apiReleased: '2026-03-26'
                    apiDocumentation: https://www.dataportal.se/rest-api-profil
                    apiStatus: development
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                internalServerError:
                  value:
                    type: https://raplp.digg.se/problems/internal-server-error
                    title: Internt serverfel
                    status: 500
                    detail: Ett ovantat fel intraffade.
                    instance: /api/v1/api-info
  /validation/validatespec:
    post:
      tags:
        - validation
      summary: Validate an OpenAPI specification.
      description: |
        Validates a base64-encoded OpenAPI specification.

        The service first performs parsing and strict validation of structural and semantic errors.
        If no strict issues are found, RAP-LP rule validation is executed for the selected rule categories.
      operationId: validateSpec
      requestBody:
        required: true
        description: Base64-encoded specification and optional validation settings.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SpecValidationRequest'
            examples:
              yamlSpecification:
                value:
                  spec: b3BlbmFwaTogMy4xLjAKaW5mbzoKICB0aXRsZTogRXhhbXBsZQogIHZlcnNpb246IDEuMC4wCnBhdGhzOiB7fQo=
                  strict: true
                  categories:
                    - DokRules
                    - UfnRules
      responses:
        '200':
          description: Validation completed without rule violations.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessfulValidationResponse'
              examples:
                success:
                  value:
                    ok: true
                    stage: rule-engine
                    format: yaml
                    payload:
                      result: []
                      report:
                        - Notering: Godkanda regler - RAP-LP
                          regler:
                            - id: DOK.03
                              område: Dokumentation
                              status: OK
        '400':
          description: The specification was invalid or violated one or more rules.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ValidationProblemDetails'
              examples:
                missingSpec:
                  value:
                    type: https://raplp.digg.se/problems/bad-request
                    title: Invalid Request
                    status: 400
                    detail: 'Required field missing: spec'
                    instance: /api/v1/validation/validatespec
                specValidation:
                  value:
                    type: https://raplp.digg.se/problems/semantic-validation
                    title: Rule validation failed
                    status: 400
                    detail: Specifikationen innehåller strukturella eller semantiska fel
                    instance: /api/v1/validation/validatespec
                    kind: spec-validation
                    stage: strict
                    format: yaml
                    issues:
                      - code: parser-error
                        message: Missing required property paths.
                        severity: error
                        path:
                          - paths
                    snippet: |
                      line 12: missing required property: paths
                ruleValidation:
                  value:
                    type: https://raplp.digg.se/problems/rule-validation
                    title: Rule validation failed
                    status: 400
                    detail: API-specifikationen bryter mot en eller flera regler enligt den svenska REST API-profilen.
                    instance: /api/v1/validation/validatespec
                    kind: rule-validation
                    stage: rule-engine
                    format: yaml
                    payload:
                      result:
                        - id: AME.04
                          område: API Message
                          krav: For faltnamn i request och response body BOR camelCase eller snake_case notation anvandas.
                          allvarlighetsgrad: WARNING
                          sökväg:
                            - components
                            - schemas
                            - Example
                            - properties
                            - problemField
                          omfattning:
                            start:
                              line: 10
                              character: 8
                            end:
                              line: 10
                              character: 20
                      report:
                        - Notering: Ej godkanda regler - RAP-LP
                          regler:
                            - id: AME.04
                              område: API Message
                              status: EJ OK
        '500':
          description: Internal server error.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                internalServerError:
                  value:
                    type: https://raplp.digg.se/problems/internal-server-error
                    title: Internt serverfel
                    status: 500
                    detail: Ett ovantat fel intraffade.
                    instance: /api/v1/validation/validatespec
...
  /validation/generate-report:
    post:
      tags:
        - validation
      summary: Generate an Excel report.
      description: Generates an Excel reconciliation report from a precomputed RAP-LP report.
      operationId: generateReport
      requestBody:
        required: true
        description: Precomputed report entries used as input when creating the Excel workbook.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GenerateReportRequest'
            examples:
              precomputedReport:
                value:
                  report:
                    - Notering: Godkanda regler - RAP-LP
                      regler:
                        - id: UFN.05
                          område: URL Format och namngivning
                          status: OK
      responses:
        '200':
          description: Excel report generated successfully.
          headers:
            Content-Disposition:
              description: Indicates attachment download and suggested filename.
              schema:
                type: string
                example: attachment; filename="avstamningsfil.xlsx"
            Content-Type:
              description: Excel media type.
              schema:
                type: string
                example: application/vnd.openxmlformats-officedocument.spreadsheetml.sheet
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        '400':
          description: Invalid report input.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                missingReport:
                  value:
                    type: https://raplp.digg.se/problems/bad-request
                    title: Invalid Request
                    status: 400
                    detail: 'Required field missing: report'
                    instance: /api/v1/validation/generate-report
        '500':
          description: Excel report generation failed.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              examples:
                generationFailed:
                  value:
                    type: https://raplp.digg.se/problems/internal-server-error
                    title: Failed to generate report
                    status: 500
                    detail: Failed to generate report.
                    instance: /api/v1/validation/generate-report
components:
  schemas:
    ValidationRules:
      type: array
      description: Available rule categories.
      items:
        $ref: '#/components/schemas/ValidationRule'
    ValidationRule:
      type: object
      additionalProperties: false
      properties:
        rule:
          type: string
          description: Rule category identifier.
          example: UfnRules
        description:
          type: string
          description: Human-readable description of the rule category.
          example: URL format och namngivning
      required:
        - rule
        - description
    SpecValidationRequest:
      type: object
      additionalProperties: false
      properties:
        spec:
          type: string
          description: Base64-encoded OpenAPI specification content.
        strict:
          type: boolean
          description: Enables strict parsing and semantic validation before RAP-LP rules are executed.
          default: true
        categories:
          type: array
          description: Optional subset of rule categories to execute.
          items:
            type: string
          examples:
            - [DokRules, UfnRules]
      required:
        - spec
...
    ApiInfo:
      type: object
      additionalProperties: false
      properties:
        apiName:
          type: string
          example: RAP-LP
        apiVersion:
          type: string
          description: Semantic version of this API.
          example: 1.0.11
        apiReleased:
          type: string
          format: date
          example: '2026-03-26'
        apiDocumentation:
          type: string
          format: uri
          example: https://www.dataportal.se/rest-api-profil
        apiStatus:
          type: string
          example: development
      required:
        - apiName
        - apiVersion
        - apiReleased
        - apiDocumentation
        - apiStatus
    ValidationResult:
      type: object
      additionalProperties: false
      properties:
        result:
          type: array
          items:
            $ref: '#/components/schemas/RapLpDiagnosticItem'
        report:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosticReport'
      required:
        - result
        - report
    SuccessfulValidationResponse:
      type: object
      additionalProperties: false
      properties:
        ok:
          type: boolean
          const: true
        stage:
          $ref: '#/components/schemas/ValidationStage'
        format:
          $ref: '#/components/schemas/SpecFormat'
        payload:
          $ref: '#/components/schemas/ValidationResult'
      required:
        - ok
        - stage
        - format
        - payload
...
    SpecFormat:
      type: string
      description: Parsed OpenAPI document format.
      enum:
        - yaml
        - json
      example: yaml
    ValidationStage:
      type: string
      description: Validation stage that produced the response.
      enum:
        - strict
        - rule-engine
      example: rule-engine
    ProblemDetails:
      type: object
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri-reference
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
          format: uri-reference
      required:
        - type
        - title
        - status
        - detail
        - instance
    ValidationProblemDetails:
      allOf:
        - $ref: '#/components/schemas/ProblemDetails'
        - type: object
          additionalProperties: true
          properties:
            kind:
              type: string
              enum:
                - spec-validation
                - rule-validation
            stage:
              $ref: '#/components/schemas/ValidationStage'
            format:
              $ref: '#/components/schemas/SpecFormat'
            issues:
              type: array
              items:
                $ref: '#/components/schemas/StrictIssue'
            snippet:
              type: string
            payload:
              $ref: '#/components/schemas/ValidationResult'
    StrictIssue:
      type: object
      additionalProperties: true
      properties:
        code:
          type: string
        message:
          type: string
        severity:
          type: string
        path:
          type: array
          items:
            type: string
      required:
        - message
    RapLpDiagnosticItem:
      type: object
      additionalProperties: true
      properties:
        id:
          type: string
          nullable: true
        område:
          type: string
          nullable: true
        krav:
          type: string
          nullable: true
        allvarlighetsgrad:
          type: string
          nullable: true
          enum:
            - HINT
            - INFORMATION
            - WARNING
            - ERROR
        sökväg:
          type: array
          items:
            type: string
          nullable: true
        omfattning:
          $ref: '#/components/schemas/Range'
        tags:
          type: array
          items:
            type: string
        relatedInformation:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosticRelatedInformation'
    DiagnosticRelatedInformation:
      type: object
      additionalProperties: false
      properties:
        location:
          $ref: '#/components/schemas/Location'
        message:
          type: string
      required:
        - location
        - message
    Location:
      type: object
      additionalProperties: false
      properties:
        uri:
          type: string
          nullable: true
        range:
          $ref: '#/components/schemas/Range'
      required:
        - range
    Range:
      type: object
      additionalProperties: false
      properties:
        start:
          $ref: '#/components/schemas/Position'
        end:
          $ref: '#/components/schemas/Position'
      required:
        - start
        - end
    Position:
      type: object
      additionalProperties: false
      properties:
        line:
          type: integer
          minimum: 0
        character:
          type: integer
          minimum: 0
      required:
        - line
        - character
...
    PopulatedDiagnosticRuleInfo:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          example: UFN.05
        område:
          type: string
          example: URL Format och namngivning
        helpUrl:
          type: string
          format: uri
        krav:
          type: string
        status:
          type: string
          example: OK
      required:
        - id
        - område
        - status
    DiagnosticReport:
      type: object
      additionalProperties: false
      properties:
        Notering:
          type: string
          example: Godkanda regler - RAP-LP
        regler:
          type: array
          items:
            $ref: '#/components/schemas/PopulatedDiagnosticRuleInfo'
      required:
        - Notering
        - regler
    GenerateReportRequest:
      type: object
      additionalProperties: true
      properties:
        report:
          type: array
          items:
            $ref: '#/components/schemas/DiagnosticReport'
      required:
        - report